fix to versioning so that the model contains sequential version numbers

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/supportadditionalneedsapi/domain/entity/ReviewScheduleHistoryEntity.kt
@@ -61,7 +61,7 @@ data class ReviewScheduleHistoryEntity(
 @Embeddable
 data class ReviewScheduleHistoryEntityKey(
   @Column
-  val version: Int,
+  val version: Long,
   @Column(name = "id")
   val id: UUID,
 )
